Splitsing van httpTeacher
