+++ frontend/src/components/teacher/RootLayoutDashboardTeacher.tsx
@@ -1,7 +1,6 @@
 import { useEffect } from "react";
 import { Outlet, useLoaderData, useSubmit } from "react-router-dom";
 import { getTokenDuration } from "../../util/teacher/authTeacher.js";
-import styles from "./RootLayoutDashboardTeacher.module.css";
 import React from "react";
 
 function RootLayoutDashboardTeacher() {
