+++ frontend/src/routes/router.tsx
@@ -114,11 +114,11 @@ export const router = createBrowserRouter([
         element: <AssignmentAdd></AssignmentAdd>,
       },
       {
-        path: 'assignments/:assignmentId',
+        path: 'assignment/:assignmentId',
         element: <Assignment></Assignment>,
       },
       {
-        path: 'assignments/:assignmentId/edit',
+        path: 'assignment/:assignmentId/edit',
         element: <AssignmentEdit></AssignmentEdit>,
       },
       {
